requirements.txt
aiohttp==3.8.1
discord.py==2.0.1
python-dotenv==0.19.2
PyYAML==6.0
